feat: add validathor + code separation
