+++ utils.ts
@@ -1,0 +1,7 @@
+export function json_safe_parse(str: string): any {
+  try {
+    return JSON.parse(str);
+  } catch {
+    return null;
+  }
+}
